Atualização Comp
